refactored monopoly view to exist in source directory as view class
