+++ index.html
@@ -2,15 +2,13 @@
 <html>
 <head>
   <meta charset="UTF-8">
-  <script src="index.js"></script>
   <script type="text/javascript" src="lib/pixi/pixi.min.js"></script>
-  <script src="bundle.js"></script>
   <title>Monopoly</title>
  </head>
  <body>
     Here's a blank html page that will execute our JavaScript!
-    <script src="monopoly_view.js"></script>
     
+    <script src="bundle.js"></script>
  </body>
  <footer>
  </footer>
